+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 0.1.3 (under development)
 
 * Move runner configuration options to runner initializer. Allow user to set the number of parallel simulations, max number of simulations, run the simulations (true/false), and verbose output.
+* Add a runner.conf file that can be used to define how the run_osw(s) behaves.
+* Add rake task to initialize the runner.conf (e.g., rake openstudio:runner:init)
 
 ## Version 0.1.2